[bugfix] handle_ready sets ready flag when ready
In blocking mode (the default), the code tight-loops waiting for self.ready to change.

self.ready only gets changed by registration succeeding in the existing codebase, which doesn't work when the sip identity doesn't need to be registered.

Change handle_ready to set self.ready, because baresip is ready once we get the baresip message for "ready".

Fixes #4